+++ src/exchange.ts
@@ -314,13 +314,13 @@ export class Exchange {
 
   private _autoPriorityFeeOffset: number = 0;
   private _autoPriorityFeeMultiplier: number = 1;
+  private _autoPriorityFeeUseMax: boolean = false;
 
   // Micro lamports per CU of fees.
   public get autoPriorityFeeUpperLimit(): number {
     return this._autoPriorityFeeUpperLimit;
   }
-  private _autoPriorityFeeUpperLimit: number =
-    constants.DEFAULT_MICRO_LAMPORTS_PER_CU_FEE;
+  private _autoPriorityFeeUpperLimit: number = constants.PRIO_FEE_UPPER_LIMIT;
 
   public get blockhashCommitment(): Commitment {
     return this._blockhashCommitment;
@@ -334,6 +334,10 @@ export class Exchange {
   public setAutoPriorityFeeScaling(offset: number = 0, multiplier: number = 1) {
     this._autoPriorityFeeMultiplier = multiplier;
     this._autoPriorityFeeOffset = offset;
+  }
+
+  public toggleAutoPriorityFeeUseMax() {
+    this._autoPriorityFeeUseMax = !this._autoPriorityFeeUseMax;
   }
 
   public updatePriorityFee(microLamportsPerCU: number) {
@@ -355,7 +359,7 @@ export class Exchange {
     if (this.isSetup) {
       throw "Exchange already setup";
     }
-    this._assets = assets.allAssets();
+    this._assets = assets.allAssets(loadConfig.network);
     this._provider = new anchor.AnchorProvider(
       loadConfig.connection,
       wallet instanceof types.DummyWallet ? null : wallet,
@@ -705,7 +709,8 @@ export class Exchange {
     return [...this._subExchanges.values()];
   }
 
-  private async updateAutoFee() {
+  // Public so you can call it as often as you want. By default gets called in the clock interval
+  public async updateAutoFee() {
     let accountList = [];
 
     // Query the most written-to accounts
@@ -730,11 +735,14 @@ export class Exchange {
         .slice(0, 20) // Grab the latest 20
         .map((obj) => obj.prioritizationFee); // Take a list of prioritizationFee values only
 
-      let median = utils.median(fees);
-      let medianScaled =
-        this._autoPriorityFeeOffset + median * this._autoPriorityFeeMultiplier;
+      let num = this._autoPriorityFeeUseMax
+        ? Math.max(...fees)
+        : utils.median(fees);
+
+      let numScaled =
+        this._autoPriorityFeeOffset + num * this._autoPriorityFeeMultiplier;
       this._priorityFee = Math.round(
-        Math.min(medianScaled, this._autoPriorityFeeUpperLimit)
+        Math.min(numScaled, this._autoPriorityFeeUpperLimit)
       );
       console.log(
         `AutoUpdate priority fee. New fee = ${this._priorityFee} microlamports per compute unit`