+++ src/risk.ts
@@ -838,21 +838,10 @@ export class RiskCalculator {
     bufferPercent: number = 5,
     maxIterations: number = 100
   ): number {
-    // Cap max leverage to 0 < maxLeverage < maxAssetLeverage
-    // Also don't cap leverage if not a taker trade, because leverage only counts positions
+    // Don't cap leverage if not a taker trade, because leverage only counts positions
     if (maxLeverage <= 0 || !isTaker) {
       maxLeverage = -1;
     }
-    if (maxLeverage != -1) {
-      let maxAssetLeverage =
-        100 /
-        convertNativeBNToDecimal(
-          Exchange.pricing.marginParameters[assets.assetToIndex(tradeAsset)]
-            .futureMarginInitial
-        );
-      maxLeverage = Math.min(maxLeverage, maxAssetLeverage);
-    }
-
     if (thresholdPercent <= 0) {
       throw Error("thresholdPercent must be > 0");
     }
@@ -951,7 +940,7 @@ export class RiskCalculator {
       2 *
       Math.max(
         0,
-        state.balance /
+        Math.max(state.balance, state.availableBalanceInitial) /
           (init * Math.min(Exchange.getMarkPrice(tradeAsset), tradePrice))
       );
     if (sizeUpperBound == 0) {
@@ -998,15 +987,12 @@ export class RiskCalculator {
         size
       );
 
-      // TODO if this is slow then do only the necessary calcs manually, there's a bunch of extra calcs in here
-      // that aren't needed in getMaxTradeSize()
       let newState = this.getCrossMarginAccountState(editedAccount);
-      let equity = newState.balance - newState.unpaidFundingTotal;
-      let nonLeverageBuffer =
-        (equity +
-          Math.min(newState.unrealizedPnlTotal, 0) -
-          newState.initialMarginTotal) /
-        equity;
+      let equity =
+        newState.balance +
+        newState.unrealizedPnlTotal +
+        newState.unpaidFundingTotal;
+      let nonLeverageBuffer = (equity - newState.initialMarginTotal) / equity;
 
       let buffer =
         maxLeverage == -1
@@ -1033,9 +1019,7 @@ export class RiskCalculator {
             10 ** constants.POSITION_PRECISION
         );
       } else if (
-        (maxLeverage == -1 &&
-          newState.initialMarginTotal >
-            equity + Math.min(newState.unrealizedPnlTotal, 0)) ||
+        (maxLeverage == -1 && newState.initialMarginTotal > equity) ||
         (maxLeverage != -1 && buffer < 0) ||
         buffer > 1
       ) {